separated it into two
